+++ packages/adapter-interfaces/src/types/index.ts
@@ -28,6 +28,7 @@ export * from "./SmsProvider";
 export * from "./ResourceServer";
 export * from "./RolePermission";
 export * from "./UserPermission";
+export * from "./UserRole";
 export * from "./Role";
 export * from "./Organization";
 export * from "./UserOrganization";