+++ packages/kysely/src/index.ts
@@ -24,8 +24,8 @@ import { createFormsAdapter } from "./forms";
 import { createResourceServersAdapter } from "./resourceServers";
 import { createRolesAdapter } from "./roles";
 import { rolePermissions } from "./role-permissions";
-import { userPermissions } from "./user-permissions";
-import { userRoles } from "./user-roles";
+import { createUserPermissionsAdapter } from "./user-permissions/adapter";
+import { createUserRolesAdapter } from "./user-roles/adapter";
 import { createOrganizationsAdapter } from "./organizations";
 import { createUserOrganizationsAdapter } from "./user-organizations";
 
@@ -53,8 +53,8 @@ export default function createAdapters(db: Kysely<Database>): DataAdapters & {
     refreshTokens: createRefreshTokensAdapter(db),
     resourceServers: createResourceServersAdapter(db),
     rolePermissions: rolePermissions(db),
-    userPermissions: userPermissions(db),
-    userRoles: userRoles(db),
+    userPermissions: createUserPermissionsAdapter(db),
+    userRoles: createUserRolesAdapter(db),
     roles: createRolesAdapter(db),
     sessions: createSessionsAdapter(db),
     tenants: createTenantsAdapter(db),