+++ packages/kysely/src/user-permissions/index.ts
@@ -1,12 +1,12 @@
 import { Kysely } from "kysely";
-import { assign } from "./assign";
+import { create } from "./create";
 import { remove } from "./remove";
 import { list } from "./list";
 import { Database } from "../db";
 
 export function userPermissions(db: Kysely<Database>) {
   return {
-    assign: assign(db),
+    create: create(db),
     remove: remove(db),
     list: list(db),
   };