+++ packages/authhero/CHANGELOG.md
@@ -1,5 +1,17 @@
 # authhero
 
+## 0.209.0
+
+### Minor Changes
+
+- d5ebc83: Update endpoints and UI for organziation memebers
+- Get organizations crud working like auth0
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.87.0
+
 ## 0.208.0
 
 ### Minor Changes